Add font_tables::read module
