+++ font-tables/src/lib.rs
@@ -2,11 +2,13 @@
 
 mod font_data;
 mod layout;
+mod read;
 mod table_ref;
 
 pub mod parse_prelude {
-    pub use crate::font_data::{Cursor, FontData, ReadError};
-    pub use crate::table_ref::{FontRead, Format, ResolveOffset, TableInfo, TableRef};
+    pub use crate::font_data::{Cursor, FontData};
+    pub use crate::read::{FontRead, Format, ReadError};
+    pub use crate::table_ref::{ResolveOffset, TableInfo, TableRef};
 
     pub use font_types::{
         BigEndian, F2Dot14, FWord, Fixed, FixedSized, LongDateTime, MajorMinor, Offset, Offset16,
